add method on PersonenkontextUpdatedEvent to check if persons is/remains Lehrer

+++ src/core/ldap/domain/ldap-event-handler.ts
@@ -151,28 +151,34 @@ export class LdapEventHandler {
             `Received PersonenkontextUpdatedEvent, personId:${event.person.id}, new personenkontexte: ${event.newKontexte.length}, deleted personenkontexte: ${event.removedKontexte.length}`,
         );
 
-        // Delete all removed personenkontexte if rollenart === LEHR
-        await Promise.allSettled(
-            event.removedKontexte
-                .filter((pk: PersonenkontextEventKontextData) => pk.rolle === RollenArt.LEHR)
-                .map(async (pk: PersonenkontextEventKontextData) => {
-                    const emailDomain: Result<string> = await this.getEmailDomainForOrganisationId(pk.orgaId);
-                    if (emailDomain.ok) {
-                        this.logger.info(`Call LdapClientService because rollenArt is LEHR, pkId: ${pk.id}`);
-                        const deletionResult: Result<PersonData> = await this.ldapClientService.deleteLehrer(
-                            event.person,
-                            emailDomain.value,
-                        );
-                        if (!deletionResult.ok) {
-                            this.logger.error(deletionResult.error.message);
+        if (!event.containsAnyCurrentPKWithRollenartLehr()) {
+            // Delete all removed personenkontexte if rollenart === LEHR
+            await Promise.allSettled(
+                event.removedKontexte
+                    .filter((pk: PersonenkontextEventKontextData) => pk.rolle === RollenArt.LEHR)
+                    .map(async (pk: PersonenkontextEventKontextData) => {
+                        const emailDomain: Result<string> = await this.getEmailDomainForOrganisationId(pk.orgaId);
+                        if (emailDomain.ok) {
+                            this.logger.info(`Call LdapClientService because rollenArt is LEHR, pkId: ${pk.id}`);
+                            const deletionResult: Result<PersonData> = await this.ldapClientService.deleteLehrer(
+                                event.person,
+                                emailDomain.value,
+                            );
+                            if (!deletionResult.ok) {
+                                this.logger.error(deletionResult.error.message);
+                            }
+                        } else {
+                            this.logger.error(
+                                `LdapClientService deleteLehrer NOT called, because organisation:${pk.orgaId} has no valid emailDomain`,
+                            );
                         }
-                    } else {
-                        this.logger.error(
-                            `LdapClientService deleteLehrer NOT called, because organisation:${pk.orgaId} has no valid emailDomain`,
-                        );
-                    }
-                }),
-        );
+                    }),
+            );
+        } else {
+            this.logger.info(
+                `Keep lehrer in LDAP, personId:${event.person.id}, because person keeps PK(s) with rollenArt LEHR`,
+            );
+        }
 
         // Create personenkontexte if rollenart === LEHR
         await Promise.allSettled(
